AMP-21861 checking currency codes to be non-empty
git-svn-id: file:///svn/branches/AMP_2_11_RELEASE@29250 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/digijava/module/aim/util/CurrencyUtil.java
@@ -779,6 +779,8 @@ public class CurrencyUtil {
 	}
 	
 	public static void deleteCurrencyRates(List<String> currencyCode) {
+		if (currencyCode.size() == 0)
+			return;
 		Query q = PersistenceManager.getSession().createQuery("delete from " + AmpCurrencyRate.class.getName() 
 				+ " o where o.fromCurrencyCode in (:currCodes) or o.toCurrencyCode in (:currCodes)");
 		q.setParameterList("currCodes", currencyCode);
